Added support for ajax real-time updates.
If track_real_time is checked (which is only possible when
end time of dataset is in the future), then ajax get requests
are done to url /ncharts/data/<id>?last_time=val
where <id> is the user's data selection, and last_time is the current
end time of their plot. The data between last_time and now
is returned to the ajax call, and the points on the highcharts
are updated, shifting over if the early times are before start_time.

Lots of additions to ncharts.js to update start_time when
tracking real-time, using the current values of time_length
and time_length_units.