Added a double check for files after a directory modification
In Dir.scan(), if the modification time of the directory has not changed,
and it has been at least 10 seconds since it was scanned due to
a modificaton, scan once more.

Without the double check, it was happening fairly frequently that a
new file was not being seen, much more frequently than would be
expected if there was fraction of a second between the mod-time
change and the appearance of the new file.