Refactor to support plotting individual soundings
Data now have one more dictionary level, the name of the series.
The series name for conventional time series is the empty string.

For a dataset of type "sounding", the series names are generated
from the time of the file, using a strftime format, currently
"%b%d_%H%MZ".

The pre-existing time series functionality seems to work. Now
to add plotting support for soundings.