Reworked handling of values in user submitted forms, improving error handling
If the user has selected a time outside of the dataset period, reset
the value for them, and send a warning via the django messages framework,
which is then displayed via the template engine on the web page.

Providing an updated value in the response form was fairly tricky, basically
had to create a new form, populated with an altered post object.